+++ public/css/error-pages.css
@@ -4,6 +4,8 @@
 @import url('/AsyncCore.github.io/public/error-pages/prism.css');
 @import url("https://fonts.googleapis.com/css?family=Bevan");
 
+/* TODO Hay qu reworkear el css para que se vea bien en todos los dispositivos y para que se vea bien en general */
+
 body {
     background: rgb(40,40,40);
     overflow: hidden;
