+++ public/error-pages/404.php
@@ -35,9 +35,10 @@
     <link color="#5bbad5" href="/img/favicon/safari-pinned-tab.svg" rel="mask-icon">
     <link href="/img/favicon/favicon.ico" rel="shortcut icon">
     <!-- CSS -->
-    <link href="/css/error-pages.css" rel="stylesheet" type="text/css">
+    <link href="/css/footer-style.css" rel="stylesheet" type="text/css">
     <link href="/error-pages/prism.css" rel="stylesheet" type="text/css">
-    <!-- JavaScript "-->"
+    <link href="/css/error-pages.css" rel="stylesheet" type="text/css">
+    <!-- JavaScript "-->
     <script defer src="/error-pages/prism.js" type="text/javascript"></script>
     <script defer src="/js/error-pages.js" type="text/javascript"></script>
     <!-- Font Awesome -->
@@ -46,32 +47,34 @@
 </head>
 <body>
 <main>
-    <div><p><span class="http">HTTP: </span><span class="num-error">404</span></p></div>
-    <div style="text-align:center;">
+    <div class="titulo"><p><span class="http">HTTP: </span><span class="num-error">404</span></p></div>
+    <div class="codigo" style="text-align:center;">
         <pre style="display:inline-block; text-align:left; white-space: pre;">
-        <code class="language-php">/* 404.php */</code>
-        <code class="language-php">&lt;?<i id="custom-php" class="jiji">php</i></code>
-            <code class="language-php">$url = "<?= $_SERVER['REQUEST_URI']?>";</code>
-            <code class="language-php">$no_sabes_escribir = false;</code>
-            <code class="language-php">$la_hemos_cagado = true;</code>
-            <code class="language-php">$encontrada = false;</code>
+                                        <code class="language-php">/* 404.php */</code>
+                                        <code class="language-php">&lt;?<i id="custom-php" class="jiji">php</i></code>
+                                            <code class="language-php">$url = "<?= $_SERVER['REQUEST_URI']?>";</code>
+                                            <code class="language-php">$no_sabes_escribir = $encontrada = false;</code>
+                                            <code class="language-php">$la_hemos_cagado = true;</code>
 
-            <code class="language-php">if ($url == !$encontrada) {</code>
-                <code class="language-php">echo "ERROR 404 - PÁGINA NO ENCONTRADA";</code>
-            <code class="language-php">}else{</code>
-                <code class="language-php">if ($no_sabes_escribir) {</code>
-                    <code class="language-php">echo "Aver hestudiao.";</code>
-                <code class="language-php">} else if ($la_hemos_cagado) {</code>
-                    <code class="language-php">echo "Lo sentimos...";</code>
-                <code class="language-php">}</code>
-            <code class="language-php">}</code>
+                                            <code class="language-php">if ($url == !$encontrada) {</code>
+                                                <code class="language-php">echo "ERROR 404 - PÁGINA NO ENCONTRADA";</code>
+                                            <code class="language-php">}else{</code>
+                                                <code class="language-php">if ($no_sabes_escribir) {</code>
+                                                    <code class="language-php">echo "Aver hestudiao.";</code>
+                                                <code class="language-php">} else if ($la_hemos_cagado) {</code>
+                                                    <code class="language-php">echo "Lo sentimos...";</code>
+                                                <code class="language-php">}</code>
+                                            <code class="language-php">}</code>
 
-            <code class="language-php">header("Location: https://www.asyncore.es/main.php", true, 302);</code>
-            <code class="language-php">exit;</code>
-        <code class="language-php">?&gt;</code>
+                                            <code class="language-php">header("Location: https://www.asyncore.es/main.php", true, 302);</code>
+                                            <code class="language-php">exit;</code>
+                                        <code class="language-php">?&gt;</code>
         </pre>
     </div>
     <a href="https://www.asyncore.es/main.php">HOME</a>
 </main>
+<?php
+    include_once '../../src/footer.php';
+?>
 </body>
 </html>
