+++ public/main.php
@@ -1,5 +1,7 @@
 <?php
+
     require __DIR__ . '/../src/utils/autoloader.php';
+    include_once __DIR__ . '/../config/databaseQueries.php';
     include_once __DIR__ . "/../src/utils/utils.php";
 
     /**
@@ -7,6 +9,11 @@
 	 * @var array  $resultado_hilos /config/databaseQueries.php
 	 * @var array  $resultado_posts /config/databaseQueries.php
 	 */
+
+    if (!isset($resultado_hilos)){
+        $resultado_hilos = [];
+        $resultado_posts = [];
+    }
 ?>
 
 <!DOCTYPE html>
@@ -110,7 +117,6 @@
 		</aside>
 		<section>
 			<?php
-                include_once __DIR__ . '/../config/databaseQueries.php';
 				$user_hilo = $resultado_hilos[0]['USERNAME'];
 				$titulo_hilo = $resultado_hilos[0]['TITULO'];
 				$contenido_hilo = $resultado_hilos[0]['CONTENIDO'];
