+++ src/LogLevels.php
@@ -1,0 +1,20 @@
+<?php
+
+    namespace src;
+
+    /**
+    * Enum con los tipos de mensajes de log.
+    * Puede ser: "INFO", "WARNING", "ERROR", "EXCEPTION" o "DEBUG".
+    * @author  Daniel Alonso Lázaro <dalonsolaz@gmail.com>
+    * @access  public
+    * @package src
+    */
+
+    enum LogLevels: string
+    {
+        case INFO = "INFO";
+        case WARNING = "WARNING";
+        case ERROR = "ERROR";
+        case EXCEPTION = "EXCEPTION";
+        case DEBUG = "DEBUG";
+    }
