+++ src/header.php
@@ -1,17 +1,12 @@
-<?php
-$success = $_POST['success'] ?? false;
-$loginEmail = $_POST['loginEmail'] ?? false;
-$usuario = $success ? substr($loginEmail, 0, strpos($loginEmail, "@")) : "Invitado";
-echo <<<HTML
 <header>
     <nav>
         <div class="logo">
             <a href="
-HTML;
-$current_path = $_SERVER['PHP_SELF'];
-$current_dir = substr($current_path, strrpos($current_path, "/"), strlen($current_path));
-$current_host = $_SERVER['HTTP_HOST'];
-$target_url = $current_dir == '/main.php' ? "https://" . $current_host . "/index.php" : "https://" . $current_host . "/main.php";
+<?php
+$currentPath = $_SERVER['PHP_SELF'];
+$currentDir = substr($currentPath, strrpos($currentPath, "/"), strlen($currentPath));
+$currentRoot = $_SERVER['HTTP_HOST'];
+$target_url = $currentDir == '/main.php' ? "https://" . $currentRoot . "/index.php" : "https://" . $currentRoot . "/main.php";
 echo $target_url;
 echo <<<HTML
                                                        ">
@@ -26,11 +21,7 @@ echo <<<HTML
         </div>
         
         <div class="user-menu">
-            <a href="/usuario-perfil.php"><button class="boton">
-HTML;
-echo $usuario;
-echo <<<HTML
-            </button></a>
+            <a href="/usuario-perfil.php"><button class="boton">Perfil</button></a>
             <div class="dropdown-content" id="dropdown">
                 <a href="/login-register.php">Login - Registro</a>
             </div>
